- Fixed issue where enemies would not display if they had zero distance to move from current position to target position.

+++ Enemy.h
@@ -5,7 +5,7 @@
 class Enemy : public Object
 {
 public:
-	Enemy(int x, int y, int targetX, int targetY, ElementType type);
+	Enemy(double x, double y, int targetX, int targetY, ElementType type);
 	~Enemy();
 
 	void InjectFrame() override;
